Update FirstCProgram
Wait do I need to comment in the code and in this box here? Maybe code comments are more for other programmers to understand the function of the program and less about my day to day struggles. But hey, in this fast-paced world of telling machines what to do and playing, uh, league of legends? Wouldn't it be nice if developers could offload some of their pent up frustration in the code of the their programs? Like Rembrandt painting "I'm sad" in all of his paintings, which he did and don't look it up just believe me it's true.